sns-auto-poster/requirements.txt
flask==3.0.3
requests==2.32.3
beautifulsoup4==4.12.3
google-cloud-firestore==2.19.0
google-cloud-aiplatform==1.71.1
vertexai==1.71.1
moviepy==1.0.3
Pillow==10.4.0
numpy==1.26.4
gunicorn==22.0.0
